ajust exception organization way; add Date (g4 not applied)

+++ src/utils/basic_function.h
@@ -30,7 +30,7 @@ std::vector<PageSlotID> Intersection(std::vector<PageSlotID> iA,
 
 Field* BuildField(const String& sRaw, FieldType iTargetFieldType);
 
-bool LeagalDate(uint16_t year, uint8_t month, uint8_t day);
+bool LeagalDate(int year, int month, int day);
 
 }  // namespace dbtrain_mysql
